server/requirements.txt
aioredis==1.2.0
aniso8601==3.0.2
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.2.1
automat==0.7.0
channels-redis==2.3.3
channels==2.1.7
constantly==15.1.0
daphne==2.2.5
django-cors-headers==2.4.0
django-filter==2.1.0
django-graphql-jwt==0.2.0
django-graphql-playground==1.0
django-pipeline==1.6.14
django==2.1.5
graphene-django==2.2.0
graphene-file-upload==1.2.2
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
namedlist==1.7
pillow==5.4.1
promise==2.2.1
pyhamcrest==1.9.0
pyjwt==1.7.1
pytz==2018.9
rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
twisted==18.9.0
txaio==18.8.1
zope.interface==4.6.0